Moved and removed some unused files
