+++ web/src/components/CreateAnnouncment.jsx
@@ -4,11 +4,11 @@ import { useNavigate } from "react-router-dom";
 //Created By Tyson
 //The Home Page dedicated to future navigation of the page 
 
-export default function Home(){
+export default function CreateAnnouncement(){
 
 return (
 
 <p>
-This is the Home
+
 </p>)
 }
