--- web/src/components/CreateShift.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
-
-//Created By Tyson
-//The Home Page dedicated to future navigation of the page 
-
-export default function CreateShift(){
-
-return (
-
-<p>
-This is the Home
-</p>)
-}
